+++ eclair-core/src/test/scala/fr/acinq/eclair/wire/ChannelCodecsSpec.scala
@@ -5,7 +5,7 @@ import java.nio.charset.Charset
 import fr.acinq.bitcoin.BinaryData
 import fr.acinq.eclair.channel.{LocalParams, RemoteParams}
 import fr.acinq.eclair.crypto.Sphinx
-import fr.acinq.eclair.randomKey
+import fr.acinq.eclair.{UInt64, randomKey}
 import fr.acinq.eclair.transactions._
 import fr.acinq.eclair.wire.ChannelCodecs._
 import fr.acinq.eclair.wire.LightningMessageCodecs._
@@ -34,7 +34,7 @@ class ChannelCodecsSpec extends FunSuite {
     val o = LocalParams(
       nodeId = randomKey.publicKey,
       dustLimitSatoshis = Random.nextInt(Int.MaxValue),
-      maxHtlcValueInFlightMsat = Random.nextInt(Int.MaxValue),
+      maxHtlcValueInFlightMsat = UInt64(Random.nextInt(Int.MaxValue)),
       channelReserveSatoshis = Random.nextInt(Int.MaxValue),
       htlcMinimumMsat = Random.nextInt(Int.MaxValue),
       toSelfDelay = Random.nextInt(Short.MaxValue),
@@ -57,7 +57,7 @@ class ChannelCodecsSpec extends FunSuite {
     val o = RemoteParams(
       nodeId = randomKey.publicKey,
       dustLimitSatoshis = Random.nextInt(Int.MaxValue),
-      maxHtlcValueInFlightMsat = Random.nextInt(Int.MaxValue),
+      maxHtlcValueInFlightMsat = UInt64(Random.nextInt(Int.MaxValue)),
       channelReserveSatoshis = Random.nextInt(Int.MaxValue),
       htlcMinimumMsat = Random.nextInt(Int.MaxValue),
       toSelfDelay = Random.nextInt(Short.MaxValue),
